Use in-house TypeScript unique symbol for PublicKey
Unfortunately, using the one from `@solana/keys` fails if the dependant library does not also install `@solana/keys` and I'd rather not import a library as a main dependency just for a type. There might be a way around this but for now let's use an in-house symbol. Updating the symbol in the future won't be a breaking change.