Define PublicKeys as base58 strings (#62)
* Represent Public Keys as base58 strings

* Fix isPublicKey and assertPublicKey

* Add string length check to assertPublicKey

* Update Signer.ts

* Update PublicKey.ts

* Update PublicKey.ts

* Refactor samePublicKey

* Refactor publicKey serializer

* Define PDAs as a tuple

* Make PublicKey and Pda types opaque

* Add SafePublicKeyInput that can be transformed without asserting

* Make PublicKey compatible with @solana/keys

* Make opaque PublicKey compatible both ways with web3.js

* Parameterize PublicKey string

* Update caniuse for browserlist

* Add changeset

* Update changesets

* Use in-house TypeScript unique symbol for PublicKey

Unfortunately, using the one from `@solana/keys` fails if the dependant library does not also install `@solana/keys` and I'd rather not import a library as a main dependency just for a type. There might be a way around this but for now let's use an in-house symbol. Updating the symbol in the future won't be a breaking change.

* Update documentation

* Update changeset